only re-render virtualized demo when in view
